Move the jazzer API to the core package
